finance_ml_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
camelot-py[cv]
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
sentence-transformers==2.2.2
plotly==5.17.0
python-multipart==0.0.6
opencv-python==4.8.1.78
ghostscript==0.7
pypdfium2==4.25.0
joblib==1.3.2
kaleido==0.2.1
transformers==4.35.2
torch==2.1.0
requests
